support delete, add tests
