+++ internal/command/sandbox/subst_test.go
@@ -1,0 +1,98 @@
+package sandbox
+
+import (
+	"bytes"
+	"encoding/json"
+	"io/ioutil"
+	"os"
+	"testing"
+)
+
+type SubstTest struct {
+	Yaml   string
+	Args   []string
+	Error  bool
+	Result string
+}
+
+// below cases hard code fields which for some reason
+// are not omitempty, although in the swagger source they are.
+// TODO: fix omitempties in go-sdk.
+var substCases = []SubstTest{
+	{
+		Yaml:   `{"name":"${dev}-service"}`,
+		Args:   []string{"dev=jane"},
+		Result: `{"endpoints":null,"name":"jane-service"}`,
+	},
+	{
+		Yaml:  `{"name":"${dev}-service"}`,
+		Args:  []string{"xdev=jane"},
+		Error: true,
+	},
+	{
+		Yaml:   `{"name":"service-${dev}"}`,
+		Args:   []string{"dev=jane"},
+		Result: `{"endpoints":null,"name":"service-jane"}`,
+	},
+	{
+		Yaml:   `{"name":"${team}-service-${dev}"}`,
+		Args:   []string{"team=gorillas", "dev=jane"},
+		Result: `{"endpoints":null,"name":"gorillas-service-jane"}`,
+	},
+	{
+		Yaml:   `{"name":"${team}${dev}"}`,
+		Args:   []string{"team=gorillas", "dev=jane"},
+		Result: `{"endpoints":null,"name":"gorillasjane"}`,
+	},
+	{
+		Yaml:   `{"name":"b${team}${dev}e"}`,
+		Args:   []string{"team=gorillas", "dev=jane"},
+		Result: `{"endpoints":null,"name":"bgorillasjanee"}`,
+	},
+	{
+		Yaml:   `{"spec":{"cluster":"${dev}-cluster"}}`,
+		Args:   []string{"dev=jane"},
+		Result: `{"endpoints":null,"spec":{"cluster":"jane-cluster","endpoints":null,"forks":null,"resources":null}}`,
+	},
+}
+
+func testSubstCase(tc *SubstTest, t *testing.T) {
+	f, err := ioutil.TempFile(".", "substCase")
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	defer os.Remove(f.Name())
+	_, err = f.Write([]byte(tc.Yaml))
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	f.Close()
+	sb, err := loadSandbox(f.Name(), tc.Args)
+	if err == nil && tc.Error {
+		t.Errorf("didn't get error on %s", tc.Yaml)
+		return
+	}
+	if err != nil {
+		if !tc.Error {
+			t.Errorf("error processing yaml %s: %s", tc.Yaml, err.Error())
+		}
+		return
+	}
+	d, e := json.Marshal(sb)
+	if e != nil {
+		t.Error(e)
+		return
+	}
+	if !bytes.Equal(d, []byte(tc.Result)) {
+		t.Errorf("unexpected got %s want %s", string(d), tc.Result)
+	}
+}
+
+func TestSubst(t *testing.T) {
+	for i := range substCases {
+		tc := &substCases[i]
+		testSubstCase(tc, t)
+	}
+}
